Made client config statically typed
This makes accessing the config values much faster, since it no longer
needs to do a lookup nor string conversion, which means we could remove
some needless copying of the values.

Overall it makes it easier to find out where settings are used and it
puts the defaults along with the declaration.

Options with default values are no longer saved to the config file. This
does not include unrecognized options, which are kept around to provide
some compatibility with older clients.

While most basic options have kept the same name, more complicated
settings like window geometry, shortcuts, outfits, etc. now have their
own XML elements. Older clients will ignore these and erase them when
saving the configuration.

+++ src/game.cpp
@@ -415,7 +415,7 @@ void Game::handleInput()
             if (setupWindow->isVisible() &&
                 keyboard.getNewKeyIndex() > KeyboardConfig::KEY_NO_VALUE)
             {
-                keyboard.setNewKey((int) event.key.keysym.sym);
+                keyboard.setNewKey(event.key.keysym.sym);
                 keyboard.callbackNewKey();
                 keyboard.setNewKeyIndex(KeyboardConfig::KEY_NO_VALUE);
                 return;
